+++ CHANGELOG.md
@@ -36,6 +36,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Update persistUpdate function to dispatch updated item after patching inventory
 - Refactor useInventory hook to improve state reference handling and avoid stale closures
 - Refactor inventory state management to use useReducer for UI state and add new actions for toggling quick add and archived item visibility
+- Refactor useInventory and inventoryReducer to manage lastFetchedAt state with useReducer
 
 ---